control/requirements.txt
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.4
Flask-Migrate==2.3.1
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
SQLAlchemy==1.2.7
Flask-Testing==0.7.1
PyJWT==1.7.1
SQLAlchemy-Utils==0.33.9
pycrypto==2.6.1
kafka-python==1.4.3
PyMySQL==0.9.3
